SSE: [Chore] Move expr service under services/
