+++ pkg/services/expr/classic/classic.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 
-	"github.com/grafana/grafana/pkg/expr/mathexp"
+	"github.com/grafana/grafana/pkg/services/expr/mathexp"
 )
 
 // ConditionsCmd is a command that supports the reduction and comparison of conditions.
