+++ pkg/services/expr/classic/classic_test.go
@@ -10,7 +10,7 @@ import (
 	ptr "github.com/xorcare/pointer"
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/grafana/grafana/pkg/expr/mathexp"
+	"github.com/grafana/grafana/pkg/services/expr/mathexp"
 )
 
 func TestConditionsCmd(t *testing.T) {
