+++ pkg/services/expr/classic/evaluator.go
@@ -3,7 +3,7 @@ package classic
 import (
 	"fmt"
 
-	"github.com/grafana/grafana/pkg/expr/mathexp"
+	"github.com/grafana/grafana/pkg/services/expr/mathexp"
 )
 
 type evaluator interface {
