+++ pkg/services/expr/classic/evaluator_test.go
@@ -3,7 +3,7 @@ package classic
 import (
 	"testing"
 
-	"github.com/grafana/grafana/pkg/expr/mathexp"
+	"github.com/grafana/grafana/pkg/services/expr/mathexp"
 	"github.com/stretchr/testify/require"
 	ptr "github.com/xorcare/pointer"
 )
