+++ pkg/services/expr/mathexp/reduce.go
