+++ pkg/services/expr/mathexp/type_series.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 
-	"github.com/grafana/grafana/pkg/expr/mathexp/parse"
+	"github.com/grafana/grafana/pkg/services/expr/mathexp/parse"
 )
 
 // seriesTypeTimeIdx is the data frame field index for the Series type's Time column.
