+++ pkg/services/expr/mathexp/types.go
@@ -3,7 +3,7 @@ package mathexp
 import (
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 
-	"github.com/grafana/grafana/pkg/expr/mathexp/parse"
+	"github.com/grafana/grafana/pkg/services/expr/mathexp/parse"
 )
 
 // Results is a container for Value interfaces.
