+++ pkg/services/ngalert/eval/eval.go
@@ -12,10 +12,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/grafana/grafana/pkg/expr"
-	"github.com/grafana/grafana/pkg/expr/classic"
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/services/datasources"
+	"github.com/grafana/grafana/pkg/services/expr"
+	"github.com/grafana/grafana/pkg/services/expr/classic"
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
 	"github.com/grafana/grafana/pkg/setting"
