+++ pkg/services/ngalert/eval/extract_md.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/grafana/grafana/pkg/expr/classic"
+	"github.com/grafana/grafana/pkg/services/expr/classic"
 )
 
 func extractEvalString(frame *data.Frame) (s string) {
